backend/requirements.txt
flask==3.0.3
flask-pymongo==2.3.0
pymongo==4.7.2
python-dotenv==1.0.1
flask-cors==4.0.1
pytesseract==0.3.13
Pillow==10.3.0
PyMuPDF==1.24.5
werkzeug==3.0.3
python-dateutil==2.9.0.post0
marshmallow==3.21.3
